Add ads to docs
