+++ src/theme/TOC/index.js
@@ -1,0 +1,11 @@
+import Toc from '@theme-original/TOC';
+import React from 'react';
+
+export default function TOCWrapper(props) {
+  return (
+    <>
+      <div class="ads-container" />
+      <Toc {...props} />
+    </>
+  );
+}
